replace UTXO with wallet.UnspentOutput
I'm kinda on the fence about this, but I think it's the right choice.
Removing the custom type makes the API work more nicely with the us
wallet packages, and reduces the size of the response. Callers will
only want to use a subset of the outputs anyway, so they can request
the info of those outputs separately. Of course, the downside is that
more API calls are required, which could strain public walrus servers.

I suspect that, if walrus actually sees any broader adoption, there
will be pressure to bloat its response objects with helpful related
information. That's fine; I expect to acquiesce in most cases. So, in
anticipation of that, it makes sense to narrow the response objects
as much as possible now, doesn't it? :)